Read du CRUD
