--- false/f6/f6a26212105dc7d91fe0f632f72533067811e8e1aac9c22f94d9014de0cede5e.php
@@ -1,93 +1,0 @@
-<?php
-
-use Twig\Environment;
-use Twig\Error\LoaderError;
-use Twig\Error\RuntimeError;
-use Twig\Extension\SandboxExtension;
-use Twig\Markup;
-use Twig\Sandbox\SecurityError;
-use Twig\Sandbox\SecurityNotAllowedTagError;
-use Twig\Sandbox\SecurityNotAllowedFilterError;
-use Twig\Sandbox\SecurityNotAllowedFunctionError;
-use Twig\Source;
-use Twig\Template;
-
-/* main/index.html.twig */
-class __TwigTemplate_b12afa7ff981d67b23d6dd6f763d1f85c4df5de9733abf52358b805a09d91b31 extends Template
-{
-    private $source;
-    private $macros = [];
-
-    public function __construct(Environment $env)
-    {
-        parent::__construct($env);
-
-        $this->source = $this->getSourceContext();
-
-        $this->blocks = [
-            'body' => [$this, 'block_body'],
-        ];
-    }
-
-    protected function doGetParent(array $context)
-    {
-        // line 1
-        return "default.html.twig";
-    }
-
-    protected function doDisplay(array $context, array $blocks = [])
-    {
-        $macros = $this->macros;
-        $this->parent = $this->loadTemplate("default.html.twig", "main/index.html.twig", 1);
-        $this->parent->display($context, array_merge($this->blocks, $blocks));
-    }
-
-    // line 3
-    public function block_body($context, array $blocks = [])
-    {
-        $macros = $this->macros;
-        // line 4
-        echo "
-    ";
-        // line 5
-        $context['_parent'] = $context;
-        $context['_seq'] = twig_ensure_traversable(($context["posts"] ?? null));
-        foreach ($context['_seq'] as $context["_key"] => $context["post"]) {
-            // line 6
-            echo "        <h2>";
-            echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["post"], "titre", [], "any", false, false, false, 6), "html", null, true);
-            echo "</h2>
-        <p>";
-            // line 7
-            echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["post"], "contenu", [], "any", false, false, false, 7), "html", null, true);
-            echo "</p>
-    ";
-        }
-        $_parent = $context['_parent'];
-        unset($context['_seq'], $context['_iterated'], $context['_key'], $context['post'], $context['_parent'], $context['loop']);
-        $context = array_intersect_key($context, $_parent) + $_parent;
-        // line 9
-        echo "
-";
-    }
-
-    public function getTemplateName()
-    {
-        return "main/index.html.twig";
-    }
-
-    public function isTraitable()
-    {
-        return false;
-    }
-
-    public function getDebugInfo()
-    {
-        return array (  70 => 9,  62 => 7,  57 => 6,  53 => 5,  50 => 4,  46 => 3,  35 => 1,);
-    }
-
-    public function getSourceContext()
-    {
-        return new Source("", "main/index.html.twig", "C:\\wamp64\\www\\Blog-PHP\\views\\main\\index.html.twig");
-    }
-}
