+++ index.php
@@ -4,8 +4,8 @@ define('ROOT', str_replace('index.php', '', $_SERVER['SCRIPT_FILENAME']));
 
 require 'vendor/autoload.php';
 
-require_once(ROOT. 'app/Model.php');
-require_once(ROOT. 'app/Controller.php');
+require_once(ROOT. 'app/AbstractModel.php');
+require_once(ROOT. 'app/AbstractController.php');
 
 // On sépare les paramètres et on les met dans le tableau $params
 $params = explode('/', $_GET['p']);
@@ -39,10 +39,10 @@ if($params[0] != ""){
 }else{
     // Ici aucun paramètre n'est défini
     // On appelle le contrôleur par défaut
-    require_once(ROOT.'controllers/Main.php');
+    require_once(ROOT.'controllers/Home.php');
 
     // On instancie le contrôleur
-    $controller = new Main();
+    $controller = new Home();
 
     // On appelle la méthode index
     $controller->index();
