+++ models/Post.php
@@ -1,6 +1,6 @@
 <?php 
 
-class Post extends Model{
+class Post extends AbstractModel{
 
     public function __construct()
     {
@@ -10,19 +10,5 @@ class Post extends Model{
         // Nous ouvrons la connexion à la base de données
         $this->getConnection();
     }
-        
-    /**
-     * Retourne un post en fonction de son id
-     *
-     * @param int $id
-     * @return void
-     */
-    public function findById(string $id){
-        $sql = "SELECT * FROM ".$this->table." WHERE `id`='".$id."'";
-        $query = $this->_connexion->prepare($sql);
-        $query->execute();
-        return $query->fetch(PDO::FETCH_ASSOC);    
-    }
-
 
 }
