+++ views/posts/index.html.twig
@@ -24,7 +24,7 @@
             <div class="post-preview col-lg-6 col-md-6 ">
                     <a href="posts/show/{{post.id}}"> 
                     <h2 class="post-title">{{ post.titre }}</h2>
-                    <h3 class="post-subtitle">Problems look mighty small from 150 miles up</h3>
+                    <h3 class="post-subtitle">{{ post.chapo }}</h3>
                 </a>
                 <p class="post-meta">
                     Posted by
