+++ views/posts/read.html.twig
@@ -1,0 +1,46 @@
+{% extends 'default.html.twig' %}
+
+
+{% block body %}
+    <h1>Commentaires index</h1>
+
+    <table class="table">
+        <thead>
+            <tr>
+                <th>Id</th>
+                <th>Titre</th>
+                <th>Chapo</th>
+                <th>Contenu</th>
+                <th>creation</th>
+                <th>Modifié</th>
+                <th>Id</th>
+                <th>Action</th>
+
+
+            </tr>
+        </thead>
+        <tbody>
+        {% for post in posts %}
+            <tr>
+                <td>{{ post.id }}</td>
+                <td>{{ post.titre | u.truncate(25, '...', false) }}</td>
+                <td>{{ post.chapo | u.truncate(25, '...', false) }}</td>
+                <td>{{ post.contenu | u.truncate(25, '...', false) }}</td>
+                <td>{{ post.creationTime }}</td>
+                <td>{{ post.updateTime }}</td>
+                <td>{{ post.id_user }}</td>
+                <td>
+                    <a href="show/{{post.id}}">show</a>
+                    <a href="edit/{{post.id}}">edit</a>
+                </td>
+            </tr>
+        {% else %}
+            <tr>
+                <td colspan="3">no records found</td>
+            </tr>
+        {% endfor %}
+        </tbody>
+    </table>
+
+    <a href="">Create new</a>
+{% endblock %}
