+++ views/posts/show.html.twig
@@ -9,11 +9,11 @@
                 <div class="col-md-10 col-lg-8 col-xl-7">
                     <div class="post-heading">
                         <h1>{{ post.titre }}</h1>
-                        <h2 class="subheading">Problems look mighty small from 150 miles up</h2>
+                        <h2 class="subheading">{{post.chapo}}</h2>
                         <span class="meta">
                             Posted by
                             <a href="#!">Start Bootstrap</a>
-                            on August 24, 2022
+                            {{post.creationTime}}
                         </span>
                     </div>
                 </div>
